Modify model/payByPrime.py app.py
